Add 'hintinputs' to provide text hints for inputs
If <label> elements are found use them, else use the value in the box
(current behaviour) or if you add 'name' to 'hintinputs' then it will
allow you to select by the name of the field.

For values that are not displayed on the input field, it adds the value
to the red hint icon.