fix: issue 916: `When zooming in Firefox 24: TypeError...'
The following revision to `firefox' (notice the `$REV_FF' in the URI):

  REV_FF=f8057a4c1e18536842dc1876e5daae9a1e58c6b7
  http://hg.mozilla.org/mozilla-central/rev/$REV_FF

altered the signature of the following [quite possibly private] method:

  _applyZoomToPref

as seen here (notice the `$REF_FF' in the URI):

  http://hg.mozilla.org/mozilla-central/rev/$REV_FF#l1.383
  -  _applyZoomToPref: function FullZoom__applyZoomToPref() {
  +  _applyZoomToPref: function FullZoom__applyZoomToPref(browser) {

I'm guessing that the response in the `vimperator' code is the following
(notice the `$REV_VIMP' in the URI):

  REV_VIMP=5862dce879ce5089c85674b49024ab1a8087de25
  http://code.google.com/p/vimperator-labs/source/detail?r=$REV_VIMP
  Fri Sep 20 02:13:03 2013 +0900
  fix: "TypeError: FullZoom._applySettingToPref is not a function"
  Fixes issue 904

However, that fix of course renders vimperator incompatible with versions
of firefox older than revision "$REV_FF"; the next officially tagged,
compatible version of firefox is:

  $ cd /path/to/firefox/mercurial/repository
  $ REVSET="first(sort(descendants($REV_FF) and tag(), date))"
  $ hg log --template '{latesttag}\n' -r "$REVSET"
  FIREFOX_AURORA_25_BASE

Hence. this changeset updates the `install.rdf' to reflect that `25.0' is
the new minimum required version for firefox.