remove pasteClipboard function
This function is mapped to <S-Insert> and seems to mimic the default
behavior (pasting whatever is in the clipboard). However, it is also the
root of issue #166. I can't think of a reason why it should exist. It
might allow the use of <S-Insert> on Mac OS X, but I don't think
vimperator is the place for that. Also, there's cmd_paste already.