Fix a bug with hints and links containing a 0
When following a link containing a "0", the status bar would always
switch to error because of an assertion on 0. This was caused by the
double usage of the "0" in hints, both the default (empty) value and the
number 0 (part of the default hintchars).

This replaces the empty value with null and adds a check so that we
don't consider the first char as a hint unless pressed after another
hint char.