Fixes and refactor in hints module (#597)
* Remove duplicated code in hints initialization and cleanup

* Move hints keys handling to hints module

* Fix a bug with hints and links containing a 0

When following a link containing a "0", the status bar would always
switch to error because of an assertion on 0. This was caused by the
double usage of the "0" in hints, both the default (empty) value and the
number 0 (part of the default hintchars).

This replaces the empty value with null and adds a check so that we
don't consider the first char as a hint unless pressed after another
hint char.

* Convert an assertion to a check in hints _checkUnique

When following a link, pressing two hint chars in a row (contained in
the link text), would return a loud error because the otherwise valid
string sequence would result in a failed assertion.

* Fix <BS> in hints module

Disconnected, can't determine since when from the history.

* Remove _canUpdate and followFirst flags in hints onEvent

_canUpdate : if we're in onEvent, something changed in the hints, so we
always update. If it's not needed, return early.
followFirst : only used with <Return>, call directly instead.

* Fix macros mode exiting on hints with letter 'q'

Fixes issue #361

* Fix hints in fullscreen

Fullscreen window is not accessible right away through window.gBrowser.
This debounces the hints generation back on the command stack so
window.gBrowser gets set.
Fixes issue #419.